refactor: 코스 기본 엔티티 클래스 수정 (#93)
