+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/entity/CourseCommentJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.course.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "course_comments")
@@ -18,5 +18,5 @@ class CourseCommentJpaEntity(
     val userId: Long,
     @Id @Tsid
     @Column(name = "course_comment_id")
-    val id: Long? = 0L,
+    override val id: Long = 0L,
 ) : BaseTimeEntity()
