+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/entity/CourseJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.course.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "courses")
@@ -18,15 +18,15 @@ class CourseJpaEntity(
     val viewCount: Long,
     @Column(columnDefinition = "text")
     val contents: String,
-    @Column(name = "secondary_region", nullable = false)
+    @Column(name = "secondary_region")
     val secondaryRegion: String,
-    @Column(name = "primary_region", nullable = false)
+    @Column(name = "primary_region")
     val primaryRegion: String,
-    @Column(name = "name", nullable = false)
+    @Column(name = "name")
     val name: String,
     @Column(name = "user_id")
     val userId: Long,
     @Id @Tsid
-    @Column(name = "course_id", nullable = false)
-    val id: Long? = 0L,
+    @Column(name = "course_id")
+    override val id: Long = 0L,
 ) : BaseTimeEntity()
