+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/entity/InterestCourseJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.course.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseEntity
 
 @Entity
 @Table(name = "interest_course")
@@ -16,5 +16,5 @@ class InterestCourseJpaEntity(
     val userId: Long,
     @Id @Tsid
     @Column(name = "interest_course_id")
-    val id: Long? = 0L,
-) : BaseTimeEntity()
+    override val id: Long = 0L,
+) : BaseEntity()
